refactor: organize dialog components based on FSD structure
