+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { BrowserRouter as Router } from "react-router-dom"
-import Header from "./widgets/ui/Header"
-import Footer from "./widgets/ui/Footer"
+import Header from "./widgets/layout/Header"
+import Footer from "./widgets/layout/Footer"
 import PostsManagerPage from "./pages/PostsManagerPage"
 
 const App = () => {
