+++ src/entities/comment/ui/index.ts
@@ -1,3 +1,4 @@
 export * from "./CommentLikeButton"
 export * from "./CommentEditButton"
 export * from "./CommentDeleteButton"
+export * from "./CommentItem"
