+++ src/widgets/comment/ui/index.ts
@@ -1,0 +1,2 @@
+export * from "./CommentAddDialog"
+export * from "./CommentEditDialog"
