+++ src/widgets/layout/index.ts
@@ -1,0 +1,2 @@
+export * from "../layout/Header"
+export * from "../layout/Footer"
