+++ src/widgets/post/PostAddDialog.tsx
@@ -1,5 +1,5 @@
+import usePosts from "@/features/post/model/usePosts"
 import { Button, Dialog, DialogContent, DialogHeader, DialogTitle, Input, Textarea } from "@/shared/ui"
-import usePosts from "../model/usePosts"
 
 /* 게시물 추가 대화상자 */
 export const PostAddDialog = () => {
