+++ src/widgets/post/index.ts
@@ -1,0 +1,4 @@
+export * from "./PostSection"
+export * from "./PostDetailDialog"
+export * from "./PostAddDialog"
+export * from "./PostEditDialog"
