--- src/widgets/ui/index.ts
@@ -1,3 +1,0 @@
-export * from "./Header"
-export * from "./CommentSection"
-export * from "./Footer"
